+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file.
 - Add `GetAgentAvailability` SDK method for auth chat
 - Pass `logger` to AMSClient
 - Add `portalContactId` in `StartChatOptionalParams` and `GetAgentAvailabilityOptionalParams`
+- Added exception on initialization failure
+- Upgraded ACSAdapter to version beta.17
+- Added `botSurveyInviteLink` and `botFormsProLocale` the `getPostChatSurveyContext()` response
 
 ### Fixed
 - Fix `ChatAdapterOptionalParams.ACSAdapter.options.egressMiddleware` being used as `ingressMiddleware`